requirements.txt
altgraph==0.17.4
colorama==0.4.6
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
pypng==0.20220715.0
pywin32==306
pywin32-ctypes==0.2.2
qrcode==7.4.2
setuptools==72.1.0
six==1.16.0
typing_extensions==4.12.2
